# AVX BF16 and single scale quant optimizations