@netrunnereve 2024-11-12
With Q4_0 and only Q4_0 we can optimize this further by adding the 16-bit products together, allowing us to do less conversions to 32-bit. With IQ4_NL and Q8_0 there's an overflow risk if our quantized weights are right at the 8-bit limit.

| model                          |       size |     params | backend    | threads |          test |                  t/s | speedup |
| ------------------------------ | ---------: | ---------: | ---------- | ------: | ------------: | -------------------: | ---: |
llama 8B Q4_0 	| 4.33 GiB |	8.03 B |	CPU |	8 |	pp512 |	9.56 ± 0.01 	| |
llama 8B Q4_0 	| 4.33 GiB |	8.03 B |	CPU |	8 |	tg128 |	7.02 ± 0.01 	| |
| llama 8B Q4_0 (PR)                 |   4.33 GiB |     8.03 B | CPU        |       8 |         pp512 |         10.65 ± 0.05 | 11% |
| llama 8B Q4_0 (PR)                 |   4.33 GiB |     8.03 B | CPU        |       8 |         tg128 |          7.70 ± 0.01 | 10% |